+++ CHANGELOG.rst
@@ -88,6 +88,8 @@ in development
   [jsjeannotte]
 * Pack on install are now assigned an owner group. The ``pack_group`` property allows to pick this
   value and default is ``st2packs``. (new feature)
+* Make sure sensor container child processes (sensor instance processes) are killed and cleaned up
+  if the sensor container is forcefully terminated (SIGKILL). (bug fix, improvement)
 
 0.13.2 - September 09, 2015
 ---------------------------